+++ src/agent/core/nxagentd.cpp
@@ -301,7 +301,7 @@ static LONG H_RestartAgent(TCHAR *pszAction, NETXMS_VALUES_LIST *pArgs, TCHAR *p
 #endif
 
 #ifdef _WIN32
-   _sntprintf(szCmdLine, 4096, _T("\"%s\" -c \"%s\" %s%s-X %lu"), szExecName,
+   _sntprintf(szCmdLine, 4096, _T("\"%s\" -c \"%s\" %s%s-X %u"), szExecName,
               g_szConfigFile, (g_dwFlags & AF_DAEMON) ? _T("-d ") : _T(""),
               (g_dwFlags & AF_HIDE_WINDOW) ? _T("-H ") : _T(""),
               (g_dwFlags & AF_DAEMON) ? 0 : GetCurrentProcessId());
@@ -339,7 +339,8 @@ static LONG H_RestartAgent(TCHAR *pszAction, NETXMS_VALUES_LIST *pArgs, TCHAR *p
    return dwResult;
 #else
    _sntprintf(szCmdLine, 4096, _T("\"%s\" -c \"%s\" %s-X %lu"), szExecName,
-              g_szConfigFile, (g_dwFlags & AF_DAEMON) ? _T("-d ") : _T(""), m_pid);
+              g_szConfigFile, (g_dwFlags & AF_DAEMON) ? _T("-d ") : _T(""),
+              (unsigned long)m_pid);
    return ExecuteCommand(szCmdLine, NULL);
 #endif
 #endif  /* _NETWARE */
